Begin building functions for field manipulation. With Tests based on test data
