+++ id_analyses/fieldmanipulation.py
@@ -4,9 +4,12 @@ Created on 16 Jun 2016
 @author: gdy32713
 '''
 
-#import h5py
-#import numpy as np
+import h5py
+import numpy as np
 #import matplotlib.pyplot as plt
+
+def maxOfData(data):
+    return max(data)
 
 if __name__ == '__main__':
     pass
