--- test/test_FieldManipulation.py
@@ -1,18 +1,0 @@
-'''
-Created on 16 Jun 2016
-
-@author: gdy32713
-'''
-import unittest
-
-
-class Test(unittest.TestCase):
-
-
-    def testName(self):
-        pass
-
-
-if __name__ == "__main__":
-    #import sys;sys.argv = ['', 'Test.testName']
-    unittest.main()
